Refactor user business logic from model to service layer
